There was a broken link created on the "edit dataset" page
- also added a test to check that the "edit dataset" page gets loaded
- cleaning a bit other tests from "package"
