+++ ckanext-hdx_theme/ckanext/hdx_theme/templates/contribute_flow/create_edit.html
@@ -736,7 +736,7 @@
                                             {% set locale = h.dump_json({'content': _('Are you sure you want to delete this dataset ? '), 'heading': _('You are deleting a dataset'), 'delete': _('Delete')}) %}
                                             <a class="btn btn-danger btn-lg"
                                                href="/api/action/hdx_dataset_purge?"
-                                               data-module-success_url="{% url_for controller='package', action='search' %}"
+                                               data-module-success_url="{% url_for 'simple_search' %}"
                                                data-module-post_data="{{ h.dump_json({'id': data.id}) }}" data-module-is_json="true"
                                                data-module="hdx_confirm-action" data-module-i18n="{{ locale }}">
                                                {{ _('Delete') }}
